+++ scripts/run_simulation.R
@@ -27,10 +27,11 @@ if (dir.exists("modeleval")) unlink("modeleval", recursive = TRUE, force = TRUE)
 dir.create("modeleval")
 
 if (file.exists("segment_states.csv")) unlink("segment_states.csv")
+if (file.exists("segment_observations.csv")) unlink("segment_observations.csv")
 
 nw <- load_gtfs("../../fulldata.db", output = "etas.pb") %>%
     realtime_feed(c(sprintf("http://localhost:3000/%s/vehicle_positions", ca[1]), 
-    # realtime_feed(c(sprintf("http://localhost:3000/%s/500/1100/vehicle_positions", ca[1]), 
+    # realtime_feed(c(sprintf("http://localhost:3000/%s/50/100/vehicle_positions", ca[1]), 
     # realtime_feed(c(sprintf("http://localhost:3000/%s/1538934818/minutes/0/vehicle_positions", ca[1]), 
                     sprintf("http://localhost:3000/%s/trip_updates", ca[1])),
                   response = "protobuf")