+++ scripts/track_simulations.R
@@ -1,5 +1,7 @@
 source("scripts/common.R")
 
+options(width = 120)
+
 library(RSQLite)
 library(dbplyr)
 
@@ -85,14 +87,22 @@ eta <- function(sim, ta, ts, sched) {
     config <- jsonlite::read_json(file.path("simulations", sim, "config.json"))
     
     etatime <- as.POSIXct(as.numeric(ts), origin = "1970-01-01")
-    etadata <- loadsim(sim, as.integer(ts))
+    etadata <- loadsim(sim, as.integer(ts)) %>% 
+        filter(trip_id == ta$trip_id[1]) %>% print %>%
+        mutate(q5 = q0.025, q95 = q0.975) %>% 
+        filter(time > timestamp)
+    
+    # names(etadata)[7] <- "q5"
+    # names(etadata)[8] <- "q95"
+
     # cat(" 3 ------\n")
     # print(etadata)
     # cat(" 4 ------\n")
     # print(unique(etadata$trip_id))
     # cat(" 5 ------\n")
     # print(ta$trip_id[1])
-    etadata <- etadata %>% filter(trip_id == ta$trip_id[1])
+    # etadata <- etadata %>% filter(trip_id == ta$trip_id[1])
+    # print(etadata)
     # cat(" 6 ------\n")
     # print(etadata %>% select(timestamp, stop_sequence, time))
     
@@ -105,17 +115,23 @@ eta <- function(sim, ta, ts, sched) {
     if (nrow(Slast) == 1)
         sched2 <- sched %>% filter(stop_sequence > Slast$stop_sequence)
     else sched2 <- sched
+    # print(etadata)
+    xl <- c(min(ta$time, sched$arrival_time), max(ta$time, sched$arrival_time))
     p <- ggplot(ta, aes(time, stop_sequence)) +
         geom_point(aes(colour = type)) +
         geom_point(aes(x = arrival_time), data = sched, color = "magenta", pch = 4) +
         geom_point(aes(x = arrival_time + delay), data = sched2, color = "magenta", pch = 2) +
         geom_vline(aes(xintercept = etatime), data = NULL, col = "red", lty = 3) + 
         ggtitle(sprintf("ETAs at %s", format(etatime, "%H:%M:%S"))) +
-        xlim(min(ta$time, sched$arrival_time), max(ta$time, sched$arrival_time))
-    # print(etadata %>% select(stop_sequence, time, q5, q100))
+        xlim(xl[1], xl[2])
+    print(etadata %>% select(stop_sequence, time, q5, q95))
     if (all(c("q5", "q95") %in% names(etadata))) {
-        p <- p + geom_segment(aes(x = astime(q0), xend = pmin(max(ta$time, sched$arrival_time), astime(q95)), yend = stop_sequence), 
-            data = etadata %>% filter(q0 > 0 & q95 > 0))
+        p <- p + geom_segment(aes(
+                x = pmax(xl[1], astime(q5)), 
+                xend = pmin(xl[2], astime(q95)), 
+                yend = stop_sequence
+            ), 
+            data = etadata %>% filter(q5 > 0 & q95 > 0))
     }
     if ("q50" %in% names(etadata)) {
         p <- p + geom_point(aes(x = astime(q50)), data = etadata %>% filter(q50 > 0), color = "black")
@@ -251,6 +267,7 @@ server <- function(input, output, session) {
     observeEvent(input$simnum, {
         ## config info
         conf <- readLines(file.path("simulations", input$simnum, "config.json"))
+        conf <- conf[!grepl("/\\*.+\\*/", conf)]
         output$simconfig <- renderPrint(jsonlite::prettify(conf))
 
         rv$hdir <- file.path("simulations", input$simnum, "history")