+++ scripts/view_network.R
@@ -7,6 +7,11 @@ get_segments <- function(f = "segment_states.csv") {
         col_types = "iinn")
 }
 
+get_data <- function(f = "segment_observations.csv") {
+    read_csv(f, col_names = c("segment_id", "timestamp", "obs_tt", "est_error"),
+        col_types = "iinn")
+}
+
 get_segment_data <- function(routes) {
     con <- dbConnect(SQLite(), "fulldata.db")
     on.exit(dbDisconnect(con))
@@ -36,8 +41,11 @@ get_segment_data <- function(routes) {
     segments
 }
 
-view_segment_states <- function(f = "segment_states.csv", segment, n = 12, speed = FALSE) {
+view_segment_states <- function(f = "segment_states.csv", o, segment, n = 12, speed = FALSE, id) {
     data <- get_segments(f) %>% filter(timestamp > 0)
+    if (!missing(id)) data <- data %>% filter(segment_id %in% id)
+    show.data <- !missing(o)
+    if (show.data) obs <- get_data(o) %>% filter(timestamp > 0)
     if (missing(segment)) {
         segment <- data %>% distinct %>% pluck("segment_id") %>% 
             table %>% sort %>% tail(n) %>% names
@@ -45,6 +53,11 @@ view_segment_states <- function(f = "segment_states.csv", segment, n = 12, speed
     data <- data %>% filter(segment_id %in% segment) %>% distinct %>%
         mutate(timestamp = as.POSIXct(timestamp, origin = "1970-01-01")) %>%
         arrange(timestamp)
+    if (show.data) {
+        obs <- obs %>% filter(segment_id %in% segment) %>% distinct %>%
+        mutate(timestamp = as.POSIXct(timestamp, origin = "1970-01-01")) %>%
+        arrange(timestamp)
+    }
 
     if (speed) {
         segdata <- get_segment_data() %>% select(road_segment_id, length)
@@ -52,17 +65,34 @@ view_segment_states <- function(f = "segment_states.csv", segment, n = 12, speed
             inner_join(segdata, by = c("segment_id" = "road_segment_id")) %>%
             mutate(speed = length / travel_time) %>%
             filter(speed < 35) %>%
-            mutate(speed = speed * 3.6, .y = speed, .e = sqrt(uncertainty * 3.6^2))
+            mutate(speed = speed * 3.6, .y = speed, .e = uncertainty * 3.6^2)
+        if (show.data)
+            obs <- obs %>%
+                inner_join(segdata, by = c("segment_id" = "road_segment_id")) %>%
+                mutate(speed = length / obs_tt) %>%
+                filter(speed < 35) %>%
+                mutate(speed = speed * 3.6, .y = speed, .e = sqrt(est_error * 3.6^2))
     } else {
-        data <- data %>% mutate(.y = travel_time, .e = sqrt(uncertainty))
+        data <- data %>% mutate(.y = travel_time, .e = uncertainty)
+        if (show.data)
+            obs <- obs %>% mutate(.y = obs_tt, .e = sqrt(est_error))
     }
+    yr <- extendrange(data$.y, f = 0.5)
     p <- ggplot(data, aes(timestamp, .y)) + 
         geom_linerange(aes(ymin = .y - .e, ymax = .y + .e),  color = 'gray') +
         geom_point() +
         xlab("Time") + 
         ylab(ifelse(speed, "Speed (km/h)", "Travel Time (seconds)")) +
         facet_wrap(~segment_id, scales = ifelse(speed, "fixed", "free_y"))
-    if (speed) p <- p + ylim(0, 100)
+    if (show.data) {
+        # yr <- extendrange(c(data$.y, obs$.y))
+        p <- p + 
+            # geom_linerange(aes(ymin = .y - .e, ymax = .y + .e),  
+            #     data = obs,
+            #     color = 'pink', lwd = 0.5) +
+            geom_point(data = obs, colour = "red", cex = 0.5)
+    }
+    if (speed) p <- p + ylim(0, 100) else p <- p + ylim(yr[1], yr[2])
     p
 }
 
@@ -81,8 +111,8 @@ map_segments <- function(f = "segment_states.csv", t = max(data$timestamp)) {
                                      speed < 55 ~ "30-55 kmh",
                                      speed < 70 ~ "55-70 kmh",
                                      TRUE ~ "70+ kmh"))
-    
-    ggplot(data, aes(intersection_lon, intersection_lat, 
+
+    p <- ggplot(data, aes(intersection_lon, intersection_lat, 
                      xend = intersection_lon_end, yend = intersection_lat_end)) +
         geom_segment(data = segdata, colour = "black", alpha = 0.05) + 
         geom_segment(aes(color = speed)) +
@@ -92,11 +122,28 @@ map_segments <- function(f = "segment_states.csv", t = max(data$timestamp)) {
         scale_colour_viridis_c(option = "B", limits = c(0, 100)) +
         facet_grid(~speed_fct) +
         xlab("") + ylab("")
+    p
 }
 
 
+view_segment_states("simulations/sim000/segment_states.csv", "simulations/sim000/segment_observations.csv", n = 20)
+view_segment_states("simulations/sim000/segment_states.csv", "simulations/sim000/segment_observations.csv", speed = TRUE, n = 20)
 view_segment_states("simulations/sim000/segment_states.csv", n = 20)
 view_segment_states("simulations/sim000/segment_states.csv", speed = TRUE, n = 20)
+
+## what is happening in segment 4691
+view_segment_states("simulations/sim000/segment_states.csv", "simulations/sim000/segment_observations.csv", n = 20, id = "4691")
+library(leaflet)
+sx <- get_segment_data() %>% filter(road_segment_id == "4691")
+
+m <- leaflet() %>%
+    addTiles() %>%
+    addMarkers(sx$intersection_lon, sx$intersection_lat, 
+        popup = "Start"
+    ) %>%
+    addMarkers(sx$intersection_lon_end, sx$intersection_lat_end,
+        popup = "End"
+    )
 
 map_segments("simulations/sim000/segment_states.csv")