+++ src/network_filter.h
@@ -1,10 +1,6 @@
 #ifndef NETWORK_FILTER_H
 #define NETWORK_FILTER_H
 
-#ifndef SIMULATION
-#define SIMULATION 0
-#endif
-
 #include "gtfs.h"