+++ src/realtime_main.cpp
@@ -1,4 +1,6 @@
 #include <vector>
+#include <iomanip>
+#include <iostream>
 
 #include "realtime_feed.h"
 
@@ -63,15 +65,13 @@ void run_realtime_model (List nw)
     // Create vehicle container
     Gtfs::vehicle_map vehicles;
 
+    Gtfs::trip_map* trips = &(gtfs.trips ());
+
     // Create parameter object
     List pars = nw["parameters"];
     Gtfs::par params (pars);
     params.print ();
-
-    // Initialize an RNG
-    std::vector<RNG> rngs (params.n_core);
-    unsigned int _seed = (unsigned int) time (0);
-    for (int i=0; i<params.n_core; ++i) rngs.at (i).set_seed (_seed++);
+    gtfs.set_parameters (params);
 
     // Allow the program to be stopped gracefully    
     signal (SIGINT, intHandler);
@@ -80,6 +80,22 @@ void run_realtime_model (List nw)
     if (params.save_timings) {
         timer.save_to ("timings.csv", "iteration,timestamp,nvehicles");
     }
+
+    // Initialize an RNG
+    std::vector<RNG> rngs (params.n_core);
+    unsigned int _seed = (unsigned int) time (0);
+    for (int i=0; i<params.n_core; ++i) rngs.at (i).set_seed (_seed++);
+
+    // Initialize network
+    #pragma omp parallel for num_threads (1)
+    for (unsigned l=0; l<gtfs.segments ().bucket_count (); ++l)
+    {
+        for (auto sl = gtfs.segments ().begin (l); sl != gtfs.segments ().end (l); ++sl)
+        {
+            sl->second.update (&params, &gtfs);
+        }
+    }
+
     int tries = 0;
     int iteration = 0;
     while (ongoing)
@@ -103,7 +119,7 @@ void run_realtime_model (List nw)
         tries = 0;
         Rcout << "\n + loaded " 
             << rtfeed.n_vehicles () << " vehicle positions and "
-            << rtfeed.n_trip_updates () << " trip updates.\n";
+            << rtfeed.n_trip_updates () << " trip updates\n";
 
         {
             std::ostringstream tinfo;
@@ -156,9 +172,12 @@ void run_realtime_model (List nw)
 #endif
         #pragma omp parallel for num_threads(params.n_core)
         for (unsigned i=0; i<vehicles.bucket_count (); ++i)
-        {       
+        {
             for (auto v = vehicles.begin (i); v != vehicles.end (i); ++v)
             {
+// #if VERBOSE > 1
+//                 if (v->second.trip ()->route ()->route_short_name () != "928") continue;
+// #endif
                 v->second.mutate (rngs.at (omp_get_thread_num ()), &gtfs);
             }
         }
@@ -171,7 +190,6 @@ void run_realtime_model (List nw)
         {
             std::ofstream fout;
             fout.open ("segment_observations.csv", std::ofstream::app);
-            // fout << "segment_id,timestamp,travel_time,uncertainty\n";
             for (auto sl = gtfs.segments ().begin (); sl != gtfs.segments ().end (); ++sl)
             {
                 if (sl->second.get_data ().size () == 0) continue;
@@ -185,9 +203,23 @@ void run_realtime_model (List nw)
                 }
             }
             fout.close ();
+
+            fout.open ("segment_states.csv", std::ofstream::app);
+            for (auto sl = gtfs.segments ().begin (); sl != gtfs.segments ().end (); ++sl)
+            {
+                if (sl->second.get_data ().size () == 0) continue;
+                fout << sl->second.segment_id () << ","
+                    << sl->second.timestamp () << ","
+                    << sl->second.travel_time () << "," 
+                    << sl->second.uncertainty () << "\n";
+            }
+            fout.close ();
         }
 #endif
 
+#if VERBOSE > 0
+        Rcout << "\n\n >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> Network Update\n";
+#endif
         // Now update the network state
         #pragma omp parallel for num_threads (1)
         for (unsigned l=0; l<gtfs.segments ().bucket_count (); ++l)
@@ -198,18 +230,59 @@ void run_realtime_model (List nw)
             }
         }
         timer.report ("updating network state");
-
+        
+#if VERBOSE > 0
+        Rcout << "\n\n >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> ETA Predictions\n";
+#endif
         
         // Predict ETAs
-        // #pragma omp parallel for num_threads(params.n_core)
-        // for (unsigned i=0; i<vehicles.bucket_count (); ++i)
-        // {
-        //     for (auto v = vehicles.begin (i); v != vehicles.end (i); ++v)
-        //     {
-        //         v->second.predict_etas (rngs.at (omp_get_thread_num ()));
-        //     }
-        // }
-        // timer.report ("predicting ETAs");
+        uint64_t curtime (rtfeed.feed()->header ().timestamp ());
+        #pragma omp parallel for num_threads(params.n_core)
+        for (unsigned i=0; i<trips->bucket_count (); ++i)
+        {
+            for (auto trip = trips->begin (i); trip != trips->end (i); ++trip)
+            {
+                if (trip->second.is_active (curtime))
+                {
+#if VERBOSE > 1
+                    if (trip->second.route ()->route_short_name () != "NX1") continue;
+#endif
+                    trip->second.update_etas (curtime, rngs.at (omp_get_thread_num ()));
+#if VERBOSE > 1
+                    trip->second.print_etas ();
+#endif
+                }
+            }
+        }
+        Rcout << "\n\n";
+        timer.report ("predicting ETAs");
+
+        // Write TRIP UPDATES to feed
+#if SIMULATION
+        std::ostringstream outputname_t;
+        outputname_t << "etas/etas";
+        if (rtfeed.feed()->has_header () && rtfeed.feed()->header ().has_timestamp ()) 
+        {
+            outputname_t << "_" << rtfeed.feed ()->header ().timestamp ();
+        }
+        outputname_t << ".pb";
+        std::string oname (outputname_t.str ());
+        write_trip_updates (trips, oname);
+#endif
+        write_trip_updates (trips, outputname);
+        timer.report ("writing ETAs to protobuf feed");
+
+//         for (unsigned i=0; i<vehicles.bucket_count (); ++i)
+//         {
+//             for (auto v = vehicles.begin (i); v != vehicles.end (i); ++v)
+//             {
+// #if VERBOSE > 1
+//                 if (v->second.trip ()->route ()->route_short_name () != "NX1") continue;
+// #endif
+//                 v->second.predict_etas (rngs.at (omp_get_thread_num ()));
+//             }
+//         }
+//         timer.report ("predicting ETAs");
 
         // Write vehicles to (new) feed
 // #if SIMULATION
@@ -225,14 +298,19 @@ void run_realtime_model (List nw)
 // #endif
 //         write_vehicles (&vehicles, outputname);
 
-//         timer.report ("writing ETAs to protobuf feed");
+        // timer.report ("writing ETAs to protobuf feed");
 
         gtfs.close_connection (true);
         timer.end ();
 
+        // std::cout << "\nPress enter to continue ...";
+        // getchar ();
         // std::this_thread::sleep_for (std::chrono::milliseconds (10 * 1000));
 
         iteration++;
+#ifdef MAXIT
+        if (iteration >= MAXIT) ongoing = 0;
+#endif
     }
 
     Rcout << "\n\n --- Finished ---\n\n";