+++ src/time.h
@@ -19,6 +19,7 @@ public:
     Time (int s);
     Time (int h, int m, int s);
     Time (std::string& t);
+    Time (uint64_t& t);
 
     // let you call Time::now () anywhere
     static Time now ();
@@ -28,6 +29,11 @@ public:
     int second () const;
     int seconds () const;
 };
+
+inline int operator-(const Time& lhs, const Time& rhs)
+{
+    return lhs.seconds () - rhs.seconds ();
+}
 
 inline bool operator==(const Time& lhs, const Time& rhs)
 {
@@ -56,7 +62,9 @@ inline bool operator>=(const Time& lhs, const Time& rhs)
 
 inline std::ostream& operator<<(std::ostream& os, const Time& t)
 {
-    os << t.hour () << ":" 
+    os 
+        << (t.hour () < 10 ? "0" : "")
+        << t.hour () << ":"
         << (t.minute () < 10 ? "0" : "")
         << t.minute () << ":" 
         << (t.second () < 10 ? "0" : "")