+++ NEWS.md
@@ -1,3 +1,10 @@
+# 0.3 - third beta release
+
+- particle filter running better
+- network running better
+- trip ETAs generated per trip using KF (instead of PF)
+
+
 # 0.2 - second beta release
 
 - real-time vehicle model running with configurable parameters